+++ package.json
@@ -16,7 +16,8 @@
     "react": "^18",
     "react-bootstrap": "^2.9.2",
     "react-dom": "^18",
-    "react-fontawesome": "^1.7.1"
+    "react-fontawesome": "^1.7.1",
+    "sass": "^1.69.7"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.2.0",
