+++ src/components/views/Score.jsx
