+++ src/pages/_app.js
@@ -1,4 +1,4 @@
-import '@/styles/globals.css'
+import '../scss/styles.scss'
 
 export default function App({ Component, pageProps }) {
   return <Component {...pageProps} />
