+++ src/scss/partials/_global.scss
