+++ src/scss/partials/_score.scss
